Extract highlight.js internal api code to its own file
